+++ service/crons/cron.go
@@ -1,4 +1,4 @@
-package scripts
+package crons
 
 import (
 	"log"
