+++ service/crons/gold_price.go
@@ -1,4 +1,4 @@
-package scripts
+package crons
 
 import (
 	"context"
